third_party/blink/web_tests/svg/dynamic-updates/SVGPatternElement-dom-patternContentUnits-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'patternContentUnits' attribute of the SVGPatternElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS patternElement.getAttribute('patternContentUnits') is null
PASS patternElement.getAttribute('patternContentUnits') is "objectBoundingBox"
PASS successfullyParsed is true

TEST COMPLETE
